tag_info差export跟import

+++ src/main/java/com/example/elandweb/dao/TagInfoRepository.java
@@ -10,7 +10,9 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface TagInfoRepository extends JpaRepository<TagInfoEntity, Long> {
+public interface TagInfoRepository extends JpaRepository<TagInfoEntity, Integer> {
         @Query(value = "SELECT * FROM tag_info LIMIT :limit OFFSET :offset  " , nativeQuery = true)
         Optional<List<TagInfoEntity>> findTagInfo(@Param("limit") int limit, @Param("offset") int offset);
+
+
 }
